backend/slr_service/requirements.txt
torch>=2.0.0
opencv-python>=4.8.0
mediapipe==0.10.14
numpy>=1.24.0
scipy>=1.10.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
groq>=0.9.0
